Adjusted for correct time display
Adjusted for correct time display in hours/days
Adjusted colour to 'warning2', this will help to highlight a GeneMod in progress compared to other activities, it is a dull orange. 
(compared to good green, injured red, training yellow, psi purple and augment grey)
Tested to be working in a local copy of the mod.